🐛 fix uint256 -> 64 in IRM & remove transfer of vaultlogic

+++ src/paperTradingCompetition/Deploy/contracts/Deploy_coordinator.sol
@@ -159,7 +159,7 @@ interface IRegistryExtended is IRegistry {
 }
 
 interface IIRMExtended is IRM {
-  function setBaseInterestRate(uint256) external;
+  function setBaseInterestRate(uint64) external;
 }
 
 interface ITokenShopExtended is ITokenShop {
